Feedback Pt. 2; directory renamed to arrayToDelimiterSeparatedList
