+++ packages/manager/src/utilities/arrayToDelimiterSeparatedList/arrayToList.ts
@@ -1,4 +1,4 @@
-export default (input: string[], separator?: string): string => {
+export default (input: string[], separator: string = ','): string => {
   if (!Array.isArray(input) || input.length === 0) {
     return '';
   }
@@ -11,10 +11,6 @@ export default (input: string[], separator?: string): string => {
     const head = input.slice(0, -1);
     const tail = input[input.length - 1];
 
-    if (separator) {
-      return `${head.join(`${separator} `)}${separator} and ${tail}`;
-    }
-
-    return `${head.join(', ')}, and ${tail}`;
+    return `${head.join(`${separator} `)}${separator} and ${tail}`;
   }
 };
